v0: move prediction to own file
function is needed elsewhere, make it available
